Sets complex data passing
Modifying the `twoWayDataBinding` with a setter function, now it can set
an property on an subobject directly, and the subobject gets a new
`twoWayDataBinding` that will set the parents field if the subobject field
is changed.

Also changing an field that is data binded to an attribute will not
change the attribute value, if the field's type is ptr, struct, array or
slice.