Add dataBindingSetter interface
For setting databindings, there where used two structs oneWayDataBinding
and twoWayDataBinding. Because twoWayDataBindings where used to set the
binded attribute of the element, but also as a setter for the property
of parent/child, it was split to two new structs.

`twoWaySetter` sets the property and `twoWayAttrSetter` sets the
attribute. Because of that the `Element`s map of two way bindings must
have also the attribute setter and the property setter. So it now
doesn't hold pointers to `twoWayDataBinding` but the new interface
`dataBindingSetter`. Also the `oneWayDataBinding` map was changed to the
new interface.