Add template element support and update README
Before this, the content of the custom element was instantiated from a
string that was passed to the `innerHTML` of the `shadowDOM`. This was
inefficient, because the browser must parse the html every time a new
custom element was created.

Now it uses the template element, whose content will be cloned to the
custom element's `shadowDOM`.

It is highly recommended to create an unexported global variable that
holds the template element, and than set the elements template in the
custom element's constructor with the `SetTemplate` function.