Fix issue with observer method invocation
When a observer method was set on a field with an complex data type
(Pointer, Struct, ...), golymer used reflect to get the method and
invoke it. But it has to run `js.Object.Interface` method on a not
wrapped object, so it crashed with a loop.

The new implementation checks just the internal js object, if it has the
method and also invokes it. (Must firts set `this` so it call the `bind`
method)