Completize the DomRepeat public api
The final decision was that DomRepeat will have an delegate element name
which will be used as the stamping element. It also has `Items` public
attribute which is set to an slice and with reflect package it will get
the values of it.

The value from the slice is set to the delegate field by name `ItemAs`.
The default name is `Item`.

User than makes changes to the slice with append or so, and indicates to
the DomRepeat the change with methods `ItemsInserted` and
`ItemsRemoved`.