Populate fields from ids in template
When the user defines an custom element in the template with an id thats
name is the same as an field in the element, golymer will automaticaly
assign field as an pointer to the element. The id must be the same as
the fields name and it's type must be the elements type.

This feature enables the user that he doesn't have to write the
assignments in the ConnectedCallback like:

`myElem.subElem = myElem.Children["subElem"].Interface().(*SubElem)`