Add routerlist, rename dom-switch attrs
dom-switch attributes are now named:
expr - the value of expretion that will be compared to cases of children
elements
case - the default name of the attribute that will be compared to the
expretion of dom-switch

Added routerlist package to the spa subpackage. In this package will be
some utils/helpers for building an single page application. RouterList
is an object that maps the `window.location.pathname`.

Fixed an issue in `attrPath.Get()`. When getting an path that doesn't
exist, the method crashed. Now it defers an function that recovers from
panic and returns `null`.